+++ feature-main/workspace-detail/src/main/java/com/seugi/workspacedetail/feature/workspacemember/WorkspaceMemberViewModel.kt
@@ -52,7 +52,7 @@ class WorkspaceMemberViewModel @Inject constructor(
         personalChatRepository.createChat(
             workspaceId = workspaceId,
             roomName = "",
-            joinUsers = listOf(targetUserId.toInt()),
+            joinUsers = listOf(targetUserId),
             chatRoomImg = "",
         ).collect {
             when (it) {
