+++ network/message/src/main/java/com/seugi/network/message/response/MessageRoomEventResponse.kt
@@ -5,24 +5,24 @@ import java.time.LocalDateTime
 
 sealed class MessageRoomEventResponse(
     @Transient open val type: String,
-    @Transient open val userId: Int,
+    @Transient open val userId: Long,
 ) {
 
     data class Raw(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
     ) : MessageRoomEventResponse(type, userId)
 
     sealed class MessageParent(
         @Transient open val timestamp: LocalDateTime,
         @Transient override val type: String,
-        @Transient override val userId: Int,
+        @Transient override val userId: Long,
     ) : MessageRoomEventResponse(type, userId) {
         data class Message(
             val id: String,
             val chatRoomId: String,
             override val type: String,
-            override val userId: Int,
+            override val userId: Long,
             val message: String,
             val messageStatus: String,
             val uuid: String?,
@@ -37,37 +37,37 @@ sealed class MessageRoomEventResponse(
 
     data class Sub(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
     ) : MessageRoomEventResponse(type, userId)
 
     data class UnSub(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
     ) : MessageRoomEventResponse(type, userId)
 
     data class DeleteMessage(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
         val messageId: String,
     ) : MessageRoomEventResponse(type, userId)
 
     data class AddEmoji(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
         val messageId: String,
         val emojiId: Int,
     ) : MessageRoomEventResponse(type, userId)
 
     data class RemoveEmoji(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
         val messageId: String,
         val emojiId: Int,
     ) : MessageRoomEventResponse(type, userId)
 
     data class TransperAdmin(
         override val type: String,
-        override val userId: Int,
+        override val userId: Long,
         val roomId: String,
         val eventList: List<Int>,
     ) : MessageRoomEventResponse(type, userId)
